# Conflicts:
#	native/cmake/predefine.cmake
#	native/cocos/audio/android/AudioPlayerProvider.cpp
#	native/cocos/audio/common/utils/include/format.h
#	native/cocos/audio/common/utils/include/minifloat.h
#	native/cocos/audio/common/utils/include/primitives.h
#	native/cocos/audio/common/utils/include/tinysndfile.h
#	native/cocos/audio/common/utils/private/private.h
#	native/cocos/bindings/manual/jsb_global.cpp
#	native/cocos/bindings/manual/jsb_module_register.cpp
#	native/external-config.json
#	scripts/build-engine/package-lock.json
#	scripts/build-engine/package.json

+++ native/cocos/editor-support/middleware-adapter.cpp
@@ -45,7 +45,7 @@ Color4F &Color4F::operator=(const Color4B &right) {
 }
 
 Color4B::Color4B() = default;
-Color4B::Color4B(uint32_t r, uint32_t g, uint32_t b, uint32_t a)
+Color4B::Color4B(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 : r(r), g(g), b(b), a(a) {}
 
 Color4B &Color4B::operator=(const Color4B &right) = default;
